+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## HEAD
+
+* Don't overwite info file if it already exists in `Tus::Storage::FileSystem` (@janko-m)
+
 ## 2.0.2 (2017-12-24)
 
 * Handle `name` and `type` metadata for Uppy compatibility (@janko-m)